third_party/WebKit/LayoutTests/fast/dom/characterdata-api-arguments-expected.txt

Tests that the CharacterData API arguments are correctly validated.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS text.data is "abcd"
PASS text.__proto__.__proto__ is CharacterData.prototype
PASS text.appendData('efg') did not throw exception.
PASS text.data is "abcdefg"
PASS text.appendData() threw exception TypeError: Failed to execute 'appendData' on 'CharacterData': 1 argument required, but only 0 present..
PASS text.data is "abcdefg"
PASS text.insertData(0, 'abcd') did not throw exception.
PASS text.data is "abcdefg"
PASS text.insertData() threw exception TypeError: Failed to execute 'insertData' on 'CharacterData': 2 arguments required, but only 0 present..
PASS text.data is "abcdefg"
PASS text.insertData(0) threw exception TypeError: Failed to execute 'insertData' on 'CharacterData': 2 arguments required, but only 1 present..
PASS text.data is "abcdefg"
PASS text.insertData(999, 'test') threw exception IndexSizeError: Failed to execute 'insertData' on 'CharacterData': The offset 999 is greater than the node's length (7)..
PASS text.data is "abcdefg"
PASS text.insertData(-4294967294, 'test') did not throw exception.
PASS text.data is "abtestcdefg"
PASS text.deleteData(4, 3) did not throw exception.
PASS text.data is "abcd"
PASS text.deleteData() threw exception TypeError: Failed to execute 'deleteData' on 'CharacterData': 2 arguments required, but only 0 present..
PASS text.data is "abcd"
PASS text.deleteData(0) threw exception TypeError: Failed to execute 'deleteData' on 'CharacterData': 2 arguments required, but only 1 present..
PASS text.data is "abcd"
PASS text.deleteData(999, 3) threw exception IndexSizeError: Failed to execute 'deleteData' on 'CharacterData': The offset 999 is greater than the node's length (4)..
PASS text.data is "abcd"
PASS text.deleteData(-1, 3) threw exception IndexSizeError: Failed to execute 'deleteData' on 'CharacterData': The offset 4294967295 is greater than the node's length (4)..
PASS text.data is "abcd"
PASS text.deleteData(-4294967294, 2) did not throw exception.
PASS text.data is "ab"
PASS text.deleteData(1, 999) did not throw exception.
PASS text.data is "a"
PASS text.replaceData(0, 0, 'abcd') did not throw exception.
PASS text.data is "abcdefg"
PASS text.replaceData() threw exception TypeError: Failed to execute 'replaceData' on 'CharacterData': 3 arguments required, but only 0 present..
PASS text.data is "abcdefg"
PASS text.replaceData(0) threw exception TypeError: Failed to execute 'replaceData' on 'CharacterData': 3 arguments required, but only 1 present..
PASS text.data is "abcdefg"
PASS text.replaceData(0, 0) threw exception TypeError: Failed to execute 'replaceData' on 'CharacterData': 3 arguments required, but only 2 present..
PASS text.data is "abcdefg"
PASS text.replaceData(999, 3, 'test') threw exception IndexSizeError: Failed to execute 'replaceData' on 'CharacterData': The offset 999 is greater than the node's length (7)..
PASS text.data is "abcdefg"
PASS text.replaceData(-1, 3, 'test') threw exception IndexSizeError: Failed to execute 'replaceData' on 'CharacterData': The offset 4294967295 is greater than the node's length (7)..
PASS text.data is "abcdefg"
PASS text.replaceData(-4294967294, 0, 'test') did not throw exception.
PASS text.data is "abtestcdefg"
PASS text.replaceData(1, 999, 'aaa') did not throw exception.
PASS text.data is "aaaa"
PASS text.substringData(4, 3) is "efg"
PASS text.substringData() threw exception TypeError: Failed to execute 'substringData' on 'CharacterData': 2 arguments required, but only 0 present..
PASS text.substringData(0) threw exception TypeError: Failed to execute 'substringData' on 'CharacterData': 2 arguments required, but only 1 present..
PASS text.substringData(4, 999) is "efg"
PASS text.substringData(4, -1) is "efg"
PASS text.substringData(-1, 2) threw exception IndexSizeError: Failed to execute 'substringData' on 'CharacterData': The offset 4294967295 is greater than the node's length (7)..
PASS text.substringData(-4294967294, 3) is "cde"
PASS successfullyParsed is true

TEST COMPLETE
